Rewrite x11key.c to simulate a dual-paddle key
I found it very hard to practise Morse using the old program. Worst of
all, the programm "swallowed" button presses, so you couldn't get in the
range of 20 WPM or higher.

The delay between pressing a button and hearing a sound was also
confusing.

The new program tries to simulate a dual-paddle key, meaning as long as
you keep the left button pressed, a series of dits will be played. If
you press the right button, it'll play a series of dahs.

I hope this is an improvement.

Also, threading has been removed since it made everything harder.